requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
pandas
openpyxl
django-filter
Pillow
gunicorn
python-dotenv
whitenoise
